+++ CHANGELOG.md
@@ -6,7 +6,9 @@ changes with their rationale when appropriate:
 ### v6.43.0.0 (uncut)
 - **http4k-ai-mcp-sdk**: [Fix] All Tool Calls in OTel are mistakenly tagged as errors.
 - **http4k-ai-mcp-sdk**: [Unlikely Break] `ServerTools` and co are now constructed via `tools()`. Simply replace the calls - the as the rest of the behaviour is unchanges.
-- **http4k-ai-mcp-sdk**: [Unlikely Break] Externalised Progress Token so that you need to pass it directly into progress calls. This makes it explicit that callers without progress tokens will not be able to track progress.
+- **http4k-ai-mcp-sdk**: [Unlikely Break] Externalised Progress Token so that you need
+- **http4k-ai-mcp-sdk**: [Unlikely Break] Drop support for JsonRpc batching as it is not supported by the MCP spec.
+  to pass it directly into progress calls. This makes it explicit that callers without progress tokens will not be able to track progress.
 
 ### v6.42.0.0
 - **http4k-***: Upgrade versions